Fix gcc and clang warnings

+++ src/binary-reader.cc
@@ -2326,13 +2326,15 @@ Result BinaryReader::ReadSections() {
         break;
       case BinarySection::Event:
         ERROR_UNLESS(options_.features.exceptions_enabled(),
-                     "invalid section code: %u", section);
+                     "invalid section code: %u",
+                     static_cast<unsigned int>(section));
         section_result = ReadEventSection(section_size);
         result |= section_result;
         break;
       case BinarySection::DataCount:
         ERROR_UNLESS(options_.features.bulk_memory_enabled(),
-                     "invalid section code: %u", section);
+                     "invalid section code: %u",
+                     static_cast<unsigned int>(section));
         section_result = ReadDataCountSection(section_size);
         result |= section_result;
         break;
